+++ odyssnet/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "2.4.0"
+__version__ = "2.5.0"
 
 from .core.network import OdyssNet
 from .training.trainer import OdyssNetTrainer